% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apiFunctions.R
\name{bindRtDeets}
\alias{bindRtDeets}
\title{Function to fetch route details, put some of them into data.frame.
NOT ESPECIALLY USEFUL, SINCE DATA IS REDUNDANT WITH visRtsToDF}
\usage{
bindRtDeets()
}
\description{
Function to fetch route details, put some of them into data.frame.
NOT ESPECIALLY USEFUL, SINCE DATA IS REDUNDANT WITH visRtsToDF
}

