% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apiFunctions.R
\name{getRouteID}
\alias{getRouteID}
\title{Function to get route ID from a route shortname}
\usage{
getRouteID(shortname)
}
\arguments{
\item{shortname}{a character giving the short name, e.g. "B34", "31"}
}
\description{
Function to get route ID from a route shortname
}

