% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apiFunctions.R
\name{nullToNa}
\alias{nullToNa}
\title{Function to get all vehicles information into a data.frame}
\usage{
nullToNa(lst)
}
\description{
Function to get all vehicles information into a data.frame
}

